feat: Add a new schema and update ClinicSchema
This commit does some changes to /models/clinic.ts: ClinicSchema
was renamed to TutorSchema and PetSchema was created for pets, as well
as interfaces for both of them.

This modification was necessary because having the 'pets' property of
TutorSchema as 'Array<Pets>' was incorrect, since 'pets' is an array of
documents of type Pets.

Some very basic validation was also added to some properties of
TutorSchema.