fix: Fix the way pets array is handled by the database
This commit fixes the problem of updatePet returning null when trying
to access a pet that was created alongside a tutor. This happened due
to the function trying to access a document that didn't exist inside
the pets collection.

The solution was to ensure that every pet inside the pets array during
the creation of a tutor is being properly added to the pets collection.